Development

+++ src/ExpenseTracker.Infrastructure/Services/ExpenseService.cs
@@ -1,11 +1,12 @@
 using ExpenseTracker.Application.Common.Interfaces.Persistence.Repositories;
 using ExpenseTracker.Application.Common.Interfaces.Services;
-using ExpenseTracker.Application.ExpenseOperations.Commands;
 using ExpenseTracker.Application.ExpenseOperations.Queries;
 using ExpenseTracker.Application.Common.Errors;
 using ExpenseTracker.Core.Entities;
 using ErrorOr;
 using ExpenseTracker.Application.ExpenseOperations.Commands.Common;
+using ExpenseTracker.Application.ExpenseOperations.Commands.Create;
+using ExpenseTracker.Application.ExpenseOperations.Commands.Update;
 
 namespace ExpenseTracker.Infrastructure.Services;
 
@@ -16,22 +17,44 @@ public class ExpenseService : IExpenseService
     {
         _expenseRepository = expenseRepository;
     }
-    public async Task<ErrorOr<int>> AddExpenseAsync(CreateExpenseCommand command, Guid userId)
+    public async Task<ErrorOr<Guid>> AddExpenseAsync(CreateExpenseCommand command, Guid userId)
     {
         Expense expense = Expense.Create(
             command.Amount,
             command.Description,
             command.CategoryId,
             userId);
-        return await _expenseRepository.AddAsync(expense);
+
+        var result = await _expenseRepository.AddAsync(expense);
+        if (result <= 0)
+        {
+            return Errors.Expense.ExpenseCreationFailed;
+        }
+        else
+        {
+            return expense.Id;
+        }
     }
 
-    public async Task<int> DeleteExpenseAsync(Guid id)
+    public async Task<ErrorOr<bool>> DeleteExpenseAsync(Guid id)
     {
-        return await _expenseRepository.DeleteAsync(id);
+        Expense expense = await _expenseRepository.GetByIdAsync(id);
+        if (expense == null)
+        {
+            return Errors.Expense.ExpenseNotFound;
+        }
+        var result = await _expenseRepository.DeleteAsync(id);
+        if (result <= 0)
+        {
+            return Errors.Expense.ExpenseDeletionFailed;
+        }
+        else
+        {
+            return true;
+        }
     }
 
-    public async Task<ErrorOr<ExpenseResult?>> GetExpenseByIdAsync(Guid id)
+    public async Task<ErrorOr<ExpenseResult>> GetExpenseByIdAsync(Guid id)
     {
         var expense = await _expenseRepository.GetExpenseByIdAsync(id);
         if (expense == null)
@@ -48,7 +71,7 @@ public class ExpenseService : IExpenseService
 
     public async Task<ErrorOr<ExpenseResult>> UpdateExpenseAsync(UpdateExpenseCommand command)
     {
-        Expense? expense = await _expenseRepository.GetByIdAsync(command.Id);
+        Expense expense = await _expenseRepository.GetByIdAsync(command.Id);
         if (expense == null)
         {
             return Errors.Expense.ExpenseNotFound;
@@ -61,6 +84,19 @@ public class ExpenseService : IExpenseService
         else
         {
             return Errors.Expense.ExpenseUpdateFailed;
+        }
+    }
+
+    public async Task<bool> CheckIfUserOwnsExpense(Guid userId, Guid expenseId)
+    {
+        Guid expenseUserId = await _expenseRepository.GetExpenseUserIdAsync(expenseId);
+        if (expenseUserId == Guid.Empty || expenseUserId != userId)
+        {
+            return false;
+        }
+        else
+        {
+            return true;
         }
     }
 }
